nonebot_plugin_group_info/pyproject.toml
[tool.poetry]
name = "nonebot_plugin_group_manager"
version = "0.1.0"
description = "NoneBot2 插件用于管理群聊信息，包括设置群名片、群名、群头像和精华消息等"
authors = ["Endless-Path <rk.chensiyu@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0"
nonebot-adapter-onebot = "^2.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"